--- src/Exceptions/TimeoutException.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace BitWasp\Stratum\Exceptions;
-
-class TimeoutException extends \Exception
-{
-
-}
